backend/requirements.txt
flask
flask-cors
requests
uvicorn
gunicorn
strawberry-graphql[flask]